edit

+++ internal/service/eth.go
@@ -5,7 +5,7 @@ import (
 )
 
 // CheckPay 检查订单是否已支付
-func (s *Service) CheckPay(ctx context.Context, id int64, address, orderNo string) error {
+func (s *Service) CheckPay(ctx context.Context, id int, address, orderNo string) error {
 	// TODO 合约逻辑
 	return nil
 }
